+++ packages/code-studio/tsconfig.json
@@ -9,74 +9,28 @@
   "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "src/**/*.jsx"],
   "exclude": ["node_modules", "src/**/*.test.*", "src/**/__mocks__/*"],
   "references": [
-    {
-      "path": "../auth-plugin"
-    },
-    {
-      "path": "../auth-core-plugins"
-    },
-    {
-      "path": "../chart"
-    },
-    {
-      "path": "../components"
-    },
-    {
-      "path": "../console"
-    },
-    {
-      "path": "../dashboard"
-    },
-    {
-      "path": "../dashboard-core-plugins"
-    },
-    {
-      "path": "../file-explorer"
-    },
-    {
-      "path": "../golden-layout"
-    },
-    {
-      "path": "../grid"
-    },
-    {
-      "path": "../iris-grid"
-    },
-    {
-      "path": "../jsapi-bootstrap"
-    },
-    {
-      "path": "../jsapi-shim"
-    },
-    {
-      "path": "../jsapi-types"
-    },
-    {
-      "path": "../jsapi-utils"
-    },
-    {
-      "path": "../log"
-    },
-    {
-      "path": "../plugin-utils"
-    },
-    {
-      "path": "../pouch-storage"
-    },
-    {
-      "path": "../react-hooks"
-    },
-    {
-      "path": "../redux"
-    },
-    {
-      "path": "../storage"
-    },
-    {
-      "path": "../utils"
-    },
-    {
-      "path": "../filters"
-    }
+    { "path": "../auth-plugin" },
+    { "path": "../auth-core-plugins" },
+    { "path": "../chart" },
+    { "path": "../components" },
+    { "path": "../console" },
+    { "path": "../dashboard" },
+    { "path": "../dashboard-core-plugins" },
+    { "path": "../file-explorer" },
+    { "path": "../golden-layout" },
+    { "path": "../grid" },
+    { "path": "../iris-grid" },
+    { "path": "../jsapi-bootstrap" },
+    { "path": "../jsapi-shim" },
+    { "path": "../jsapi-types" },
+    { "path": "../jsapi-utils" },
+    { "path": "../log" },
+    { "path": "../plugin-utils" },
+    { "path": "../pouch-storage" },
+    { "path": "../react-hooks" },
+    { "path": "../redux" },
+    { "path": "../storage" },
+    { "path": "../utils" },
+    { "path": "../filters" }
   ]
 }
