+++ packages/redux/src/actions.ts
@@ -2,6 +2,7 @@ import type { Action } from 'redux';
 import type { ThunkAction } from 'redux-thunk';
 import type { CommandHistoryStorage } from '@deephaven/console';
 import type { FileStorage } from '@deephaven/file-explorer';
+import type { DeephavenPluginModuleMap } from '@deephaven/plugin-utils';
 import {
   SET_PLUGINS,
   SET_USER,
@@ -13,7 +14,6 @@ import {
   SET_SERVER_CONFIG_VALUES,
 } from './actionTypes';
 import type {
-  DeephavenPluginModuleMap,
   RootState,
   ServerConfigValues,
   User,
